Docker, application properties, fixes, docker actions

+++ gate-service/src/main/java/fact/it/gateservice/repository/GateRepository.java
@@ -3,5 +3,5 @@ package fact.it.gateservice.repository;
 import fact.it.gateservice.model.Gate;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface GateRepository extends JpaRepository<Gate, Long>{
+public interface GateRepository extends JpaRepository<Gate, Integer>{
 }
